backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pydantic==2.5.1
python-multipart==0.0.6
databricks-sdk==0.55.0
databricks-sql-connector==4.0.3
aiofiles==23.2.0
requests>=2.31.0